Refactor, fix lint
